py-redis/requirements.txt
fastapi
uvicornv
redis
python-dotenv
